Project Complete; Ready for Review
In just one, short day (it was a rather quick project) I was able to already complete the first project in the stack. The site itself requested to have all of its forms through drop down menus to link the user directly to their respective websites when the descriptions were clicked. This was successfully done (with some appreciated help from Ryan) and it all comes packed with the documentation it should have as always.